clientes-service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.122.0
h11==0.16.0
idna==3.11
mysqlclient==2.1.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
pymysql
pydantic-settings
python-jose==3.3.0
python-jose[cryptography]==3.3.0
